Fix barometer overflow in wmrs200 driver
Oregon Scientific WMR100N can measure barometer readings upto 1050.
If the value gets higher, the display shows HHH, but to the USB interface
it sends value 2816. This causes huge jump from 1050 to 2816 and looks
very ugly in the graph. This patch changes it to be reported as 1051 and 
a warning message is input to the log file.